Hp Bar complete

+++ Assets/Scripts/InGame/GameObject/Tower/Arrow.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Arrow : MonoBehaviour
 {
-    public float damage = 25.0f;
+    public int damage = 25;
     public GameObject m_target = null;                      //타겟
     public Vector3 targetPosition = Vector3.zero;           //타겟의 위치
